refactor main
