+++ .eslintrc.js
@@ -37,6 +37,7 @@ module.exports = {
         "@typescript-eslint/member-ordering": "error",
         "@typescript-eslint/no-explicit-any": "warn",
         "@typescript-eslint/no-require-imports": "error",
+        "@typescript-eslint/no-non-null-assertion": "off",
         "@typescript-eslint/ban-ts-ignore": "off",
         "@typescript-eslint/no-unused-vars": ["error", {
             "varsIgnorePattern": "^_"
