+++ .gitignore
@@ -8,4 +8,4 @@ release/
 coverage
 *.log
 package-lock.json
-test-level*.db
+*.db
